+++ apps/router-e2e/__e2e__/headless/app/(test)/about.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import { Text } from 'react-native';
-
-export default function Page() {
-  return <Text testID="content">About</Text>;
-}